Add `add-history` and `remove-history` commands.
The new Readline commands have optional hook functions to let the host
hook the add or remove operations.  I wanted the commands to be added
outside of Readline, but its use of `rl_last_func` made that very
problematic:  the `remove-history` command needs to behave differently
depending on whether a `history-search-backward/forward` is active.  It
was a more isolated change to add two commands than to redesign how
either `rl_last_func` works or how the history search modality works.

The clink host now maintains a vector of line_id for the history entries
it reads from the history banks.  This enables it to translate from
Readline's history index to line_id (bank and file offset) to update the
history banks according to any history removals made by Readline.

This works because the history banks mark entries as removed, for
obvious performance and cross-process synchronization reasons.  That
means line_id are immutable even across processes (unless you manually
alter the history file, but don't do that, and anyway it seems that
clink merges and rereads history banks often).