Entering a directory changes to it.
If the line is . or .. or ends in / or \ and is a directory, then change
to the directory and return a blank line to the host.

(Incidentally, CMD.EXE was originally meant to have that behavior!  But
due to an uninitialized variable, it never worked correctly.  For those
who may have access to the CMD.EXE source code, look in cext.cpp for
SFE_ISDIR inside SearchForExecutable and see if you can spot the bug.)