Readline now optionally supports backslash.
Setting the variable rl_backslash_path_sel to non-zero causes backslash
to be accepted as a path separator character.  This affects everything,
including the default word break characters list.

Setting the variable rl_preferred_path_separator causes filename
completion to use that as the path separator when displaying and
inserting matches.

Additionally there are two new functions:
- rl_is_path_separator(c) returns whether c is a path separator, based
on the value of rl_backslash_path_sep.
- rl_last_path_separator(string) returns a pointer to the last path
separator in string, or NULL if there are no path separators.  This is
similar to strrchr(s, '/') except that it
allows cross platform support.