Fix truncation and memory corruption in str<>.
The reserve() method reserved room for exactly the specific characters,
not including a nul terminator.  However, for example, load() in
settings.cpp set data[size] = '\0' which can write past the end of the
allocated buffer.  In practice the overrun will only affect runtime
behavior when size is a multiple of 64, because of how reserve() pads to
a multiple of 64.

It's safer to make reserve() include room for a nul terminator.  The new
exact flag can be used to reserve the specified (without adding room for
a nul terminator or padding to a multiple of 64).