Fixed mishandled unbound keys.
In all versions of clink, the terminal handling sends escape sequences
for keys like Home, Up, PageDown, etc.  Both clink (binding_resolver)
and Readline map these terminal escape sequences as chords.

But that's a problem because leading characters in the chord may be
leading characters for other chords, but once a character isn't part of
any bound chord, the rest of the chord is treated as normal input.

That's why pressing certain key combinations inputs stray characters.
Even 0.4.8 has that problem, but it's harder to see because all of the
Home, Up, PageDown, etc terminal sequences are the same length, so any
unbound keys are the same length as other bound keys.  But other key
sequences do encounter the problem even in 0.4.8.

The fix I settled on is to make a key_tester interface that a
terminal_in can call to test if a terminal key sequence is bound by
the host.  If it's not bound then the key sequence is just discarded.