Add a match type.
Readline performs quite a lot of redundant disk IO for stat() calls to
check the file type for completion matches.  This makes printing the
list of matches slow.

Also, it isn't correct to apply filename completion handling to things
that aren't filenames.

So this change introduces keeping track of the type for each match.
This will enable avoiding stat() calls and more accurate processing for
completions.