Fix hook trampoline on Windows Insider builds.
Not compatible with Windows Insider Build 20150
https://github.com/mridgers/clink/issues/543

The write_trampoline_out function used a hard-coded location relative
to the API it patched, and it verified that the hard-coded location
contained only NOP/INT3 instructions.

In recent Windows Insider builds, the hard-coded location no longer
contains NOP/INT3 instructions, but there are plenty of nearby runs
that are large enough to write the patched code into.

This change now scans backwards nearby, looking for a large enough run
of NOP/INT3 instructions to safely patch.