Fix hang when using Alt+H.
A latent bug in the printer class accidentally tried to emit the nul
terminator as though it were part of the string.  That was exposed by
the change to pass output through to the console if it advertises VT
emulation, due to a latent bug in the console output wrapper function.