Fix completion de-duplication.
For example, "foo" (directory) and "foo" (doskey alias) should not be treated
as duplicates because one completes as "foo\" and the other completes as
"foo ".  They are functionally different entries, and both should be
kept.  The new match_type stuff makes this differentiation possible to
do efficiently.  Also this change sets a specific sort order for
different match types.