The dll_cache was wasting my time.
A final build only copied the dll to the dll_cache if the cache was
empty.  But I've had to debug final builds often while tracking down
hook and injection bugs, and I keep forgetting that I have to empty the
cache otherwise it'll use an older version of the dll.

So now it also compares modtime and file size of the origin and cached
files.