Default to locale sort for completions.
Readline was using strcmp and _stricmp to sort completions.  That means
it was using ANSI routines to sort UTF8 strings (which will produce
incorrect results), it was sorting without locale awareness, and
filenames were sorted differently than Windows sorts them.

Now it converts from UTF8 to UTF16 and uses CompareStringW with the
user's default locale.

I was initially concerned about performance due to the conversions
within the sort comparator, but it turns out the char versions of the
Windows string comparison functions all internally convert to UTF16
before comparing, so the performance should be essentially the same.