Stop restricting pathish vs non-pathish matches.
It breaks having cmd commands + doskey aliases + files in the same list.
Really it needs to only restrict when the typed prefix has a path
separator, but currently the match pipeline doesn't seem to reliably
know the typed prefix.

It was just a check to prevent match generators from creating a
weird/confusing state for the user.  For now the generators need to be
well-behaved.