Make pause() consistent about where it breaks.
It was breaking after pause() if the debugger hadn't started yet, but
breaking *inside* pause() if the debugger was already started.

Breaking after pause() is preferable because then the stack and source
are informative.  Breaking inside forces the user to run at least one
command to show a stack trace and/or step to get out of pause().