Fix the history test.
It reads using a pathologically small buffer in order to test edge
cases.  But that split the CTAG and lost track of eating the CTAG,
exposing it to the test, which didn't like that at all.

Now the line iterator will thoroughly eat the CTAG line no matter how
small the read buffer is.