Fixed performance displaying matches.
Displaying matches was slow because Readline writes everything one byte
at a time, which incurs significant processing overhead across several
layers.

Added a `rl_completion_display_matches_func` variable and a
`display_matches()` function to resolve the performance satisfactorily.

It's now potentially possible to revert the changes to feed Readline
match type information.  It's only used when displaying matches, and
when inserting a match to decide whether to append a path separator.
Maybe could just add a callback for inserting.  But it's not
straightforward since Readline sorts the matches, making it difficult
to translate the sorted index to the unsorted list held by
`matches_impl`.