Fixed truncated message.
Readline has an off-by-one mistake when using vsnprintf.
It adds room for the null terminator on growing the message buffer, but
tells vsnprintf the size minus the null terminator.  And modern
vsnprintf has a security feature that forcibly null terminates rather
than risking buffer overread.

The smallest fix was to add room for one more byte, but a better fix
would be to correct the arithmetic, and to examine all other vsnprintf
usage for similar mistakes.