Classifying commands has a problem:
A line can have multiple commands, e.g.:
"cmd1 arg1 & cmd2 arg1 | cmd3 arg1"

In that case, the new stop_at_cursor logic was parsing the entire line
as a single command.

Now it will parse only the command containing the cursor.

That's still not right, though -- it needs to iteratively parse all of
the commands and e.g. classify as "cacaca".