Regenerate config.h, and fix wchar_t.
I noticed that config.h hadn't been regenerated for Readline 8.0 and
that some of the integration was a little fragile, so I regenerated
config.h and cleaned up RL_LIBRARY_VERSION.

Along the way I discovered that Readline expects that wchar_t is 4
bytes, but in Visual Studio it's 2 bytes.  So that led to abstracting
wchar_t, mbrtowc, and wcrtomb.

Though it didn't seem to affect issue #26 (cursor position when
surrogate pairs are present).