Fixed building color sequences for Readline.
The attribute class accidentally always had the "set" flags set for all
fields, so when a color was "default" it got treated as though it was
non-bold and non-underline.

That made hidden files show up in the normal color, rather than in
whatever color would have been used if the file weren't hidden.