Match wildcards with menu-complete.
The menu-complete family of commands (including clink-popup-complete)
now supports wildcards.

The match.wild setting controls whether wildcard support is enabled.

Also leading periods are skipped when matching the filename component of
a path, provided the filename component of the pattern has a wildcard.
This allows "bu" to match ".build" like CMD (and Win32 in general) does.