Rewrite script load locations.
Clink v1.x is too incompatible with v0.x with respect to how it loads
scripts.  This particularly affects Cmder, but also affects general
usage.

The new order attempts to find a balance between backward compatibility
and sensible behavior.

It starts with the undocumented `--scripts` flag that exists only for
Cmder.  Next it loads from the `clink.path` setting.  If that's not set,
it falls back to the profile directory (from `--profile` or
`%USERPROFILE%\AppData\Local\clink`).  Finally it always tries to load
from `%CLINK_PATH%` if it's set.

As far as I can tell, this achieves the behavior required by Cmder.