Rerun match pipeline when display filter active.
Generators used to run at the moment of completion and/or displaying
matches.  In v1.x and higher, match generators run at the beginning of a
word and the full set of matches is remembered and dynamically filtered
as typing continues.  The way match display filters used to work, they
(and an associated generator) typically relied on being run at the
moment they were needed.

So for backward compatibility it's necessary to rerun match generators
right before invoking a match display filter.

This should ensure the performance impact is only incurred when a match
display filter is actually active.