Fix #33; completing quoted paths
Readline only quotes if what it's actually inserting needs quoting.
So it doesn't quote the lcd.

On Windows that's cumbersome, e.g. `\pro` completing to `\Program` is
annoying because you can't just type a space, you have to manually
insert a leading quote character.

This change quotes the lcd (longest common prefix) if the next character
of any match would need to be quoted.