Synthesize %HOME% when not set.
Readline and various Lua scripts expect %HOME% to be set.
Clink v0.4.x internally set %HOME% when not set.

This change hooks getting environment variables and if HOME is queried
and doesn't exist then it synthesizes it from HOMEDRIVE and HOMEPATH,
or uses USERPROFILE if those aren't set either.

This fixes issue #38.