Use native terminal mode when %WT_SESSION% present
Although the existence of %WT_SESSION% does not reliably mean the
current CMD instance is hosted inside Windows Terminal, it will
generally be true since it's unusual to for example `start cmd.exe` from
inside a Windows Terminal session.  And even if that does happen, the
ability to run Windows Terminal suggests that the OS supports native
terminal emulation mode.