Fixed Readline bug inserting dir matches.
Readline has a bug when inserting a dir match in the middle of the line
(vs at the end of the line).  It had a special case check to avoid
inserting a double slash at the end of the line, but the check needs to
be broader:  whenever not at the beginning of the line.  Otherwise it
turns `\win_  foo` (cursor at `_`) into `\Windows\\_  foo`.