Adjust script load order.
The cmder-powerline-prompt scripts have a config file named
_powerline_config.lua.  In the file system, that sorts AFTER the other
scripts, which undermines being a config file.  It's really a bug in
cmder-powerline-prompt.  But it's kind of too late to rename the file,
and it's kind of reasonable to assume/wish that '.' and '_' files load
first.

So now Clink adjusts the script load order within each directory.  First
it loads all scripts starting with '.' or '_', and then it loads the
rest of the scripts.

This is actually pretty nice.  So I do hope it doesn't lead to an
opposite compatibility problem.

This also finally introduces a str_base subclass (str_moveable) with a
move constructor so it can be used in e.g. std::vector.