Fix #48; menu-complete gets stuck
`menu-complete` gets stuck when new match dir is same length as previous
match dir.  The regression was introduced as part of the
`match_pipeline` in the v1.0.0 rewrite.  It only considered the word
offset/length and didn't examine the word content.  So it didn't
consider `clink\lib\` different from `clink\lua\`.