Fix match generation optimization.
It may be possible to optimize key comparison to avoid unnecessary
generate passes, but it's very easy to accidentally break other cases
where its necessary to trigger a generate pass.

Remember that argmatchers and generators may need to parse the same word
differently based on its position or on preceding input line content.