Fixed Cmder startup.
It relies on clink.lua getting loaded, but Clink no longer uses such a
file, and for compatibility Clink should ignore clink.lua in all script
path directories.

So to avoid requiring any change in Cmder, Clink checks if a clink.lua
file exists in the first script path and loads it.