Fixed to not expand doskey alias if not at beginning.
When `doskey.enhanced` is disabled, only the first word on the line
is expanded if it matches a doskey alias, unless preceded by a space,
just like in CMD.

When it's enabled, the first word of each command delimited by `&` or `|` is expanded, but it ignored spaces.  That made it impossible to bypass a doskey alias.

Now it respects a space prefix to prevent expanding doskey aliases.  But
it allows one space after a `&` or `|`, so it takes two spaces to
disabled doskey alias expansion after the first command.

It may be klunky to explain, but it enables all of `aa & aa & aa` to
expand, and none of ` aa &  aa &  aa` to expand.  This feels more
intuitive to me, from a visual perspective.