Fixed #64; wrong cursor position when PROMPT contains BEL
This was functioning as intended and documented; bash behaves the same.
However, Clink automatically adds the documented `\001` and `\002`
characters around unprintable ANSI escape codes, and it makes sense to
also automatically handle the audible BEL (^G) character.

Clink documentation is also updated to be clear that for any other
unprintable characters, the PROMPT author is responsible for marking
them.