Fix `foo -x:` argmatcher syntax.
Clink 0.4.9 accidentally interpreted the next part as the next arg
position.  But that's because Clink and Readline got out of sync about
which word was being completed.

Clink 1.x didn't do any completion.  Although that was actually the
intended behavior, I want to strike a balance between compatibility and
reasonable behavior.

So now Clink recognizes the `-x:` flag syntax and does file completion,
and also colors the whole argument as a flag, regardless what follows
the `:` or `=` character.  To generate other matches, a lua script can
register a generator and provide code to handle the case.

This fixes issues #61 and #62.