Fixed #66; crash when injecting into 32 bit cmd
Turning on exception support for the whole project in v1.1.21 broke the
remote thread injection code, but only for 32 bit final release builds.

Now exception support is off by default for the whole project, is forced
off for the remote thread injection code, and is only turned on for the
terminal code that actually needed it in order to support regular
expressions in `console.findprevline()` and `console.findnextline()`.