Fix #76; parse directory token more carefully.
`cd ../../` was getting misinterpreted because it was being passed
directly to GetFileAttributesW, which correctly parses it into an empty
string internally.

Clink must parse the input line to ensure it is a single token before
passing it to GetFileAttributesW.  And some commands such as `cd` defeat
the directory shortcut feature (if not quoted) even if there's a path
that contains them (such as an atual path named "cd\foo").