Fix #79; International keyboard layout problem.
Oops.  To detect the Space key it isn't enough to check if the virtual
key code is VK_SPACE.  It's also necessary to check if the character
code is 0x20 (space) or 0x00 (special key).

The United States - International keyboard layout can produce input
where the virtual key code is VK_SPACE but the character code is a
different character.  This happens when inputing one of the accent
characters followed by Space.  The Space key ends up inserting the
previously typed accent character by itself, so the virtual key code is
VK_SPACE but the character code is not 0x20 and is not 0x00.