Fix `rl.setmatches()` when line is not empty.
Multiple places reset the Lua stack to 0 as a way of cleaning up.
I might have introduced all of them, or maybe I copied an incorrect
pattern from somewhere, I'm not sure.

In any case, now there is a helper class to save/restore the stack top
and all instances of resetting the top have been updated to use the new
helper class.