The completion commands can now complete `~` by itself.
The `expand-tilde` inputrc config variable is `on` by default in Clink.
When it's on, the completion commands are now able to expand `~` by
itself, in addition to expanding `~\etc`.  However, they still do not
expand `~etc` (no path separator) because that syntax refers to user
names rather than file names, which is less useful on Windows.