Fix #99; press UP key will get last command from other tabs
Clink is designed to store one master history list, and for all
instances of Clink to use that master history list.  I don't want to
make Clink try to guess which consoles should use which history lists.

However, this change adds a `%CLINK_HISTORY_LABEL%` envvar that can
qualify the master history file name, allowing an instance of Clink to
be configured to use a different master history list from other
instances.  Changing the envvar takes effect at the next input prompt,
and it can be changed as many times as desired during a session.

All of the usual history list behaviors continue to work exactly the
same as before, apart from the ability to change which master history
file is being used.