Detect possible antivirus interference.
If Clink takes more than 5 seconds to be injected, show a message that
it's taking a long time.

If Clink takes more than 30 seconds to be injected, stop waiting and
show a message that it timed out, and suggesting to consider adding an
exclusion for Clink in the antivirus software.

If the wait itself fails, write diagnostic info to the log file.