Fixed input keys handled by OS after certain Lua code.
If Lua code uses `io.popen()` and returns without finishing reading the
output, then the process may complete after editing has already resumed.
In that case the console input mode may be incorrect, and may allow the
OS to intercept key sequences that Clink should handle.

Debug builds of Clink detected and asserted about the situation, so now
Clink detects and corrects the situation as best as it can.