Fix potential script errors when coroutine is canceled.
When a coroutine was canceled because it was from the wrong generation,
it wasn't actually erroring out.  It just returned a nil file with an
error message.  And that caused the calling script to fail.

Now it actually errors out, aborting further execution of the coroutine.
And it suppress the error message since it's not the script's fault.