Fix #132; Prompt is not visible.
A math mistake counted lines that Readline doesn't display, causing the
prompt to walk up the screen when changing the terminal width, if the
prompt takes up more than one line.

The reason it can mess up in ConsoleZ even during workspace reload is
that ConsoleZ tends to trigger a bunch of extra "window resize" events
even when the window is not actually resizing.