Fix #119; reuse registered parsers for doskey macros.
If a command doesn't have an argmatcher but is a doskey macro, Clink now
automatically expands the doskey alias and looks for an argmatcher for
the expanded command.

This enables aliases like `gco=git checkout $*` to automatically reuse
a `git` argmatcher and produce completions for its `checkout` argument.

However, it only expands the doskey alias up to the first `$`.  So
complex aliases like `foo=app 2$gnul text $*` or `foo=$2 $1` might
behave strangely.