Fix quirk inherited from Readline.
When displaying matches, Readline prints spaces to pad EVERY column,
even the last column.  To work around that issue, Readline decrements
the number of columns it will print if the padding would reach the
edge of the screen.

Clink doesn't print padding on the last column.  So Clink doesn't need
that workaround and can display all the columns that fit.