Fix right side prompt with certain term caps.
Readline uses some terminal capabilities that can shift characters left
or right in a line.  So it's necessary to clear the right side prompt if
those term caps are used on the first line and the right side prompt is
showing.

Also added the "ch" term cap to set the horizontal cursor position.