Fix crash in test after the host shuffle.
It's great that things no longer have an inverted dependency on
host.cpp.  But this has a side effect:  now global variables from it
(such as settings) are not present.  That caused the tests to crash.

It looks like the only setting that didn't really belong in host.cpp was
"prompt.async", and it is now in rl_module.cpp.