Hooking works with MinGW now.
The crash was because MinGW and MSVC manage the stack differently.
A pointer to a local variable `replace` was added to a Detours
transaction.  The transaction updated the variable's storage space
after the local scope had ended.  MSVC didn't encounter a problem
(which to be clear was pure luck), but with MinGW that storage space
ended up containing a return address, and the stack got corrupted.

Which, by the way, explains why the stack looked impossible.