Fix `rl.getpromptinfo().inputlinecount`.
Readline cleared `_rl_vis_botlin` when the input line was accepted.
Callers may need that value even after the Readline editor has ended.

E.g. to be able to erase the prompt afterwards, or find the prompt on
the screen in order to use other APIs to extract the actual prompt text
from the screen.