Replace the implementation of `os.clock()`.
The built-in Lua implementation of `os.clock()` uses the C runtime's
`clock()` function.  The MSVC implementation is documented as only
working for the first 24 days a program is running.

That caused async prompt filtering to stop working after 24 days.

The new implementation stays accurate even if the program continues to
run for many years.  As a bonus side effect, it also has microsecond
precision.