Fix `rl_buffer:getcursor()`.
It had returned one less than the actual cursor position.

No one reported the bug, so I expect fixing it will not be experienced
as a breaking change.  The bug is explained in the documentation,
including the version number range where the bug existed, so it's
possible for scripts to be backward compatible with older versions of
Clink if they want to be.