Fix test to not accidentally request a VT response.
When a certain test failed, the error output accidentally contained an
embedded VT sequence to request terminal attributes.  And so the
terminal synthesized input text in response.  Oops.

On the other hand, this means the combination of `clink.print()` and
`console.readinput()` can be potentially interesting for interacting
directly with the terminal.