Fix single line prompt equal to the width of the screen.
Two bugs in `expand_prompt()`.

BUG 1:

Prior to RL 8.0 `expand_prompt()` missed returning the number of
invisible characters in NIFLP if the prompt wrapped because it equaled
the width of the screen.  That caused the cursor to be positioned
incorrectly for the next few characters in the input line.

RL 8.0 seems to have solved a different issue where printing escape
codes could get split at the right edge of the screen, causing them to
be printed as plain text.

BUG 2:

In RL 8.0 `expand_prompt()` counts a prompt line break when the physical
characters reaches the screen width.  But `update_line()` relies on
escape codes being included on the previous line, so that physical
characters will wrap the cursor to the next line.  This caused the
cursor to end up on the wrong line.