Fixed #187 differently.
- `doskey.enhanced` is `true` by default again, but if a macro contains
`$*` or `$1`...`$9` tags within quotes then `|` and `&` symbols are part
of the arguments, not command separators (and expansion stops after that
macro).
- Fixed compatibility when `doskey.enhanced` is `false`;
`alias|whatever` should not expand `alias` because there is no space
after it.
- Fixed doskey expansion when `>&` redirection symbol is present;
`command >& alias` should not attempt to expand `alias` because it's a
redirection file name, not a command.