Fix #191; `cd /d -` isn't supported.
The parsing of `cd -` was overly simplistic.  Now it handles subtleties
like tabs, optional spaces, the `/d` flag, and `cd/d` adjacent to each
other.

It won't try to parse compound "if" statements or command separators
(`|` or `&`), as that gets very complicated and could introduce
incompatibilities, and is not something that one typically types as
input.