Fix #193; now `history.max_lines` again supports unlimited.
The 50K limit has been removed, and 0 again means "unlimited".

(In reality there is a 999999 limit, but in practical terms it's
essentially unlimited.  It also doesn't support deleting items past
512 MB or so in the history file.  But if one has that large a history
file, there are bigger problems such as very long delays when showing
each prompt, because of needing to reload the history file.)