Fixed `color.message` setting.
Readline has some fragile heuristics for figuring out when to clear to
the end of the line after printing the local prompt string.

The implementation for `color.message` ran afoul of that fragility
(regression introduced in v1.1.11).  Using the font face stuff works
around that here.

Readline has pre-existing problems in those heuristics and I won't
attempt a broader fix at this time.  I think the mode strings from the
inputrc file are also susceptible to the same problems.