Fix loading settings defined by Lua scripts.
The first time Lua scripts are loaded, settings get loaded and their
values are correctly visible to Lua scripts while the Lua scripts are
being loaded.

But any time Lua gets RE-loaded (not the first load), then the Lua
scripts see the _default_ values of their settings.  The correct values
are visible after the Lua scripts are loaded.

So it's a niche case, but the behavior was definitely incorrect.