Fix history loss when `history.shared` changes.
If a Clink session started with `history.shared` true and with
`history.dupe_mode` "erase_prev", and later `history.shared` became
false, then any commands run failed to get added to the history.  But
they did get removed from the history.

REPRO:
- Start Clink while `history.shared` is `true` and `history.dupe_mode` is "erase_prev".
- Run `clink set history.shared false`.
- Press [Up] a few times to choose a command from history.
- Press [Enter].
EXPECTED:  The chosen command is removed from the history and then added
to the end of the history.
ACTUAL:  The chosen command is removed from the history (it is not added
to the end).