Add `terminal.east_asian_ambiguous` setting.
This controls how the width is determined for the East Asian Ambiguous
range of characters (which includes the lambda character), and now
gives users the ability to override the width for ambiguous characters.

Also fixed a bug where workaround to use the current font to measure the
character width returned 2 the first time each character was
encountered in each edit line session.