Add `match.preview_rows` setting.
When there are "too many" matches, the `clink-select-complete` command
normally prompts whether to show matches just like the `complete`
command does.

When `match.preview_rows` is set, then `clink-select-complete` instead
shows the specified number of rows of matches as a preview when there
are more rows.  The list is expanded to show all rows when the selection
moves past the preview rows.  This is similar to completion in the fish
shell.