Always use raw pointers from matches array.
The `do_popup_list` function returned a copy of the chosen match instead
of the actual match.

This is in preparation for maybe being able to use a match pointer as a
unique key to look up the match type and other attributes.  The broader
goal is to be able to eventually remove the embedded match types while
still allowing duplicate matches.