Added colors for executable-ness of command word.
When `color.unrecognized` is set, then if the command word is not
recognized it is displayed in the specified color.

When `color.executable` is set, then if the command word is an
executable file it is displayed in the specified color.

The lookups are performed on a background thread to keep typing
responsive.

Invalid or remote drives are not searched, because they can take up to 2
minutes to time out.  If a single background thread were used, then long
timeouts would cause starvation.  If multiple background threads were
used, then it would cause scale/cost problems.