The default_settings file overrides the defaults.
Any settings in the default_settings file override the built-in defaults
for settings.  So `clink set foo.bar clear` first tries to reset foo.bar
to whatever is in default_settings, and then falls back to the built-in
default.

This holds true when saving settings as well:

Any settings whose value matches whatever is in default_settings will be
omitted from the clink_settings file.  If there is no corresponding
default value in default_settings, then if the value matches the
built-in default then the setting will be omitted from clink_settings.

This means you can share a default_settings file and it fully replaces
the built-in defaults.