Improved default_settings file:
Looks for default settings in two places, in this order:
- {profile_dir}\default_settings- {binaries_dir}\default_settings

On save, settings are written if they were set to a value explicitly, or
if they were cleared to the built-in default value (being cleared to a
custom default value writes out the setting).

This means now a single profile can be shared by multiple binary
directories without losing setting values.  This is a niche scenario and
is probably only relevant when actively modifying the Clink project.