Fix buffer overrun in History library.
It assumes `\` is never the last character in the line, and blindly
skips the next character.

In the CMD shell, history probably shouldn't even be interpreting
backslash as an escaping character in the first place.  Which probably
has further implications on some of the quote handling.

In particular, `pgm \dir\ word` will probably be considered as TWO
arguments instead of three, and that will need to be fixed.