Fix #231; installer fails when uncheck "Use versioned directory".
The uninstaller invokes a temporary uninstaller exe, and returns before
the temporary uninstaller exe actually finishes.

There is a `_?=` flag that forces it to invoke the uninstaller directly,
without making a temporary exe.  But you have to pass in the uninstall
path in order to do that.  The best way to know the correct uninstall
path for the old version is to get parent directory from the
uninstaller's exe file regkey.

A bit messy and hard to discover, but at least it works.