`clink-select-complete` handles mouse clicks and wheel.
Fixed passing unhandled binding with params upstream; param parsing
needed to be reset.

However, passing a binding with params upstream will not work correctly
if the upstream module does not support the key binding.  This is
because win_terminal_in avoids producing key bindings that can't be
handled, but it can't predict when the key binding might be passed
upstream.  So anything that uses result.pass() is responsible for making
sure it's going to work out ok.