Focus events are for internal OS use.
MSDN warns to ignore them.  And without a reliable way to test whether
the console currently has focus, it wouldn't be useful/functional for
Clink to respond to it anyway.

Clink would need to know the current focus state so it could avoid
adjusting Quick Edit mode.  Otherwise even if it restores the Quick Edit
mode appropriately in response to losing focus, it will adjust it again
the next time it waits for input.