Optimize: `collect_words` can cache doskey aliases.
When collecting words for many lines at once, the doskey alias lookup
accounted for a large chunk of the total CPU time.

It was 35% of the total CPU time in a debug build, where around 25-40%
of the total time was spent capturing callstacks for heap allocations.
In a release build the CPU time would have likely been almost 50%.

The cache reduced the time from 35% to 15%.  The effectiveness of the
cache is inversely proportional to the uniqueness of commands in the
history.