Work around a Readline hack with shadowed keys.
Readline dispatches shadowed keys by temporarily altering the keymap
entry for a different key, and then dispatches that.

So if the command that gets invoked happens to traverse the keymap (such
as Clink's `clink-show-help` or Readline's `dump-functions`) then it
sees an inaccurate keymap and reports that key binding incorrectly.

Use a temporary variable, rather than altering the global keymap.