Fix heap corruption.
Starting in v1.3.1:
Match generation holds on to a freed pointer, and continues to corrupt
a specific block of memory for the rest of the session.

Starting in v1.3.3:
When `color.executable` or `color.unrecognized` are set, then showing a
new prompt line holds on to a freed pointer and continues to corrupt a
specific block of memory for the rest of the session.

Starting in v1.3.9:
If there are any deprecated argmatchers (i.e. using v0.4.9 APIs for
creating argmatchers), then reloading Lua (Ctrl+X,Ctrl+R) holds on to a
freed pointer and continues to corrupt a specific block of memory for
the rest of the session.