Added back support for merging argmatchers.
- v1.0.0 lost support for merging argmatchers.
- v1.1.4 restored very limited support for merging argmatchers.
- v1.3.11 restores support for merging argmatchers.

The merge is still simple.  It is still far from perfect.  But it should
be sufficient for common cases.  When merging linked argmatchers, it
still only merges the first argument position of the old and new linked
argmatchers.

Also, this fixes some bugs in the v0.4.9 merging implementation (the
merging unit tests for v0.4.9 don't pass in v0.4.9).

And further, when using the new (v1.0.0+) argmatcher syntax, this now
supports some additional merge cases that the old v0.4.9 argmatcher
syntax couldn't support.