Added `clink.oncommand()`.
This registers a function to be called whenever a command is entered in
the edit line.

Entering a directory is not a command.  Entering anything else may be a
command, but it may be "executable" or "unrecognized".

The motivation behind adding this was to enable a Lua script to look up
information about a command and take action accordingly.

For example, a Lua script could look up a fish completion script for the
command, and build an argmatcher (only) when it's needed.