Prior argmatchers can influence match generation.
Previously, only the command containing the cursor was parsed during
match generation.

The new `onarg` callback requires also parsing all commands prior to the
command containing the cursor, so that the callbacks have an opportunity
to influence match generation for the command containing the cursor.

For example, `pushd c:\foo\ & type a`TAB will generate completions for
`c:\foo\a*`, instead of `.\a*`.