Enable Readline's sigint handling.
Clink can't respond to Ctrl+C without losing the ability to handle a
bunch of key bindings.  But it can at least respond to Ctrl+Break while
displaying matches.

However, ideally Clink would also be able to respond to Ctrl+Break while
generating matches.  That's much more difficult because of Lua match
generator integration.  That is not present yet, and it's unclear
whether it can be possible, or to what degree.