Update CMD.EXE's environment snapshot.
Readline sets the LINES and COLUMNS environment variables.  But the
`set` command in CMD.EXE looks at a snapshot of the environment, so the
updates aren't visible in CMD.EXE until the `set` command is used to set
or clear a variable.

So, now Clink forces CMD.EXE to update its environment snapshot at the
same time it queries the ERRORLEVEL variable.

Also, `os.setenv()` was setting the variable twice, and has been
cleaned up to only set it once.