Minor breaking change in keyboard driver.
Try to handle Alt-Ctrl-[, Alt-Ctrl-], Alt-Ctrl-\ better, at least in
keyboard layouts where the [, ], or \ is the regular (unshifted) name
of the key.

This changes the key sequence for these, to make them be the regular
sequences one would expect on other platforms.

However, when `terminal.raw_esc` is false then the '[' is still handled
as before, to avoid generating an ambiguous key sequence that's a prefix
of other key sequences.