Add `unicode.iter()`.
Remove `unicode.sub()` because its coordinate system is difficult to
use and it offers limited value.

The new `unicode.iter()` function returns an iterator that steps through
the input string one codepoint at a time.  Each iteration returns the
string for the codepoint, the integer value of the codepoint, and a
boolean indicating whether the codepoint is a combining mark.

This provides an approach that is both more flexible and more efficient.