Fix `file:close()` with `io.popenyield()`.
When `io.popen()` or `io.popenyield()` is used in a coroutine, then the
returned `file:close()` intentionally didn't wait for the process to
exit.

That was so garbage collection's implicit close call wouldn't block.
But it had the side effect of making explicit `file:close()` unable to
return the actual status from the spawned process.

Now explicit `file:close()` calls yield until the process exits, and
then they return the actual status from the spawned process, just as they
would in the main (co)routine.